# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wealth Levels API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /builds:
    get:
      operationId: listBuilds
      tags: [builds]
      summary: List user builds (income streams / businesses)
      responses:
        "200":
          description: List of builds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Build"
        "401":
          description: Unauthorized
    post:
      operationId: createBuild
      tags: [builds]
      summary: Create a new build
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildInput"
      responses:
        "201":
          description: Created build
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Build"
        "401":
          description: Unauthorized
  /builds/{id}:
    patch:
      operationId: updateBuild
      tags: [builds]
      summary: Update a build
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildUpdate"
      responses:
        "200":
          description: Updated build
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Build"
        "401":
          description: Unauthorized
        "404":
          description: Build not found
    delete:
      operationId: deleteBuild
      tags: [builds]
      summary: Delete a build
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Build not found

  # ─── Budget ──────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get a specific user's full data (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: User not found
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: User not found
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: User not found
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get platform-wide stats (admin only)
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/leaderboard:
    get:
      operationId: adminGetLeaderboard
      tags: [admin]
      summary: Get top users by level and net worth (admin only)
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
components:
  schemas:
...
    CharacterStats:
      type: object
      required: [str, vit, int, agi, per, luk, unspentPoints]
      properties:
        str:           { type: integer, description: "Investment Growth" }
        vit:           { type: integer, description: "Net Worth Stability" }
        int:           { type: integer, description: "Diversification" }
        agi:           { type: integer, description: "Savings Rate" }
        per:           { type: integer, description: "Budget Discipline" }
        luk:           { type: integer, description: "Emergency Fund" }
        unspentPoints: { type: integer }
    Dashboard:
      type: object
      required: [id, userId, displayName, title, level, xp, xpToNext, netWorth, totalAssets, stats, systemLog]
      properties:
        id:          { type: integer }
        userId:      { type: integer }
        displayName: { type: string }
        title:       { type: string }
        level:       { type: integer }
        rank:        { type: string }
        xp:          { type: integer }
        xpToNext:    { type: integer }
        netWorth:    { type: number }
        totalAssets: { type: number }
        stats:
          $ref: "#/components/schemas/CharacterStats"
        systemLog:
          type: array
          items: { type: string }
...
    Quest:
      type: object
      required: [id, userId, title, category, targetAmount, currentAmount, xpReward, frequency, completed, createdAt]
      properties:
        id:            { type: integer }
        userId:        { type: integer }
        title:         { type: string }
        description:   { type: ["string", "null"] }
        category:      { type: string, enum: [SYSTEM, SELF] }
        targetAmount:  { type: ["number", "null"] }
        currentAmount: { type: number }
        xpReward:      { type: integer }
        frequency:     { type: string, enum: [DAILY, WEEKLY, MONTHLY, ONGOING] }
        completed:     { type: boolean }
        createdAt:     { type: string, format: date-time }
...
    Skill:
      type: object
      required: [id, userId, name, category, type, level, streakCount, lastCheckin, createdAt]
      properties:
        id:          { type: integer }
        userId:      { type: integer }
        name:        { type: string }
        category:    { type: string, enum: [INVESTMENT, SAVINGS, KNOWLEDGE] }
        type:        { type: string, enum: [SYSTEM, SELF] }
        level:       { type: integer }
        streakCount: { type: integer }
        lastCheckin: { type: ["string", "null"], format: date-time }
        xpToNext:    { type: integer }
        createdAt:   { type: string, format: date-time }
...
    Build:
      type: object
      required: [id, userId, name, description, rank, revenue, expenses, createdAt]
      properties:
        id:          { type: integer }
        userId:      { type: integer }
        name:        { type: string }
        description: { type: string }
        rank:        { type: string, enum: [S, A, B, C, D, E] }
        revenue:     { type: number }
        expenses:    { type: number }
        netProfit:   { type: number }
        createdAt:   { type: string, format: date-time }
    BuildInput:
      type: object
      required: [name, description]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
        rank:        { type: string, enum: [S, A, B, C, D, E] }
        revenue:     { type: number }
        expenses:    { type: number }
    BuildUpdate:
      type: object
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: string }
        rank:        { type: string, enum: [S, A, B, C, D, E] }
        revenue:     { type: number }
        expenses:    { type: number }

    # ─── Budget ────────────────────────────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, clerkId, displayName, isAdmin, level, netWorth, xp, questCount, skillCount, buildCount, createdAt]
      properties:
        id:          { type: integer }
        clerkId:     { type: string }
        displayName: { type: string }
        avatarUrl:   { type: ["string", "null"] }
        isAdmin:     { type: boolean }
        level:       { type: integer }
        rank:        { type: string }
        netWorth:    { type: number }
        xp:          { type: integer }
        questCount:  { type: integer }
        skillCount:  { type: integer }
        buildCount:  { type: integer }
        createdAt:   { type: string, format: date-time }
    AdminUserDetail:
      type: object
      required: [user, dashboard, quests, skills, builds]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        dashboard:
          $ref: "#/components/schemas/Dashboard"
        quests:
          type: array
          items:
            $ref: "#/components/schemas/Quest"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        builds:
          type: array
          items:
            $ref: "#/components/schemas/Build"
    AdminUserUpdate:
      type: object
      properties:
        displayName: { type: string }
        isAdmin:     { type: boolean }
    AdminStats:
      type: object
      required: [totalUsers, activeUsers, avgLevel, avgNetWorth, totalQuests, totalBuilds]
      properties:
        totalUsers:   { type: integer }
        activeUsers:  { type: integer }
        avgLevel:     { type: number }
        avgNetWorth:  { type: number }
        totalQuests:  { type: integer }
        totalBuilds:  { type: integer }
    LeaderboardEntry:
      type: object
      required: [rank, userId, displayName, level, netWorth, xp]
      properties:
        rank:        { type: integer }
        userId:      { type: integer }
        displayName: { type: string }
        avatarUrl:   { type: ["string", "null"] }
        level:       { type: integer }
        tierRank:    { type: string }
        netWorth:    { type: number }
        xp:          { type: integer }