# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Wealth Levels API
 * OpenAPI spec version: 0.1.0
 */
import type { SkillCategory } from './skillCategory';
import type { SkillType } from './skillType';

export interface Skill {
  id: number;
  userId: number;
  name: string;
  category: SkillCategory;
  type: SkillType;
  level: number;
  streakCount: number;
  /** @nullable */
  lastCheckin: Date | null;
  xpToNext?: number;
  createdAt: Date;
}